checkpoint: Moved Medias
