+++ musicbrainz_db_lite/src/lib.rs
@@ -16,10 +16,10 @@ pub use crate::models::musicbrainz::artist_credit::ArtistCredits;
 pub use crate::models::musicbrainz::genre::Genre;
 pub use crate::models::musicbrainz::label::Label;
 pub use crate::models::musicbrainz::main_entities::MainEntity;
+pub use crate::models::musicbrainz::media::Media;
 pub use crate::models::musicbrainz::recording::Recording;
 pub use crate::models::musicbrainz::relations::Relation;
 pub use crate::models::musicbrainz::release::LabelInfo;
-pub use crate::models::musicbrainz::release::Media;
 pub use crate::models::musicbrainz::release::Release;
 pub use crate::models::musicbrainz::release_group::ReleaseGroup;
 pub use crate::models::musicbrainz::tags::Tag;
