+++ musicbrainz_db_lite/src/models/musicbrainz/mod.rs
@@ -1,3 +1,4 @@
+pub mod media;
 use std::sync::Arc;
 
 use crate::DBClient;
